Add --cleanup and --media parameters.
Look at stream duration when possible instead of format duration. The
difference is relevant when re-running softchat with an existing media
file without manually removing subtitles from a previous run.

This doesn't solve the problem of the video stream ending before the subtitles,
but it bounds the problem to the maximum length of the subtitles
animation.